Implementation des routes

+++ src/main/java/com/chatop/model/Message.java
@@ -17,13 +17,13 @@ public class Message {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
+	private Integer id;
 	
 	@Column(name = "rental_id")
-	private Long rentalId;
+	private Integer rentalId;
 	
 	@Column(name = "user_id")
-	private Long userId;
+	private Integer userId;
 	
 	private String message;
